ist eine datei, kein ordner

+++ Vps/Util/ClearCache.php
@@ -194,9 +194,9 @@ class Vps_Util_ClearCache
         } else {
             $dir = new DirectoryIterator($path);
             foreach ($dir as $fileinfo) {
-                if ($fileinfo->isFile()) {
+                if ($fileinfo->isFile() && $fileinfo->getFilename() != '.gitignore') {
                     unlink($fileinfo->getPathName());
-                } elseif (!$fileinfo->isDot() && $fileinfo->isDir() && $fileinfo->getFilename() != '.svn' && $fileinfo->getFilename() != '.gitignore') {
+                } elseif (!$fileinfo->isDot() && $fileinfo->isDir() && $fileinfo->getFilename() != '.svn') {
                     $this->_removeDirContents($fileinfo->getPathName(), $server);
                     @rmdir($fileinfo->getPathName());
                 }
